+++ modules/pepper-mouse_admin/src/main.c
@@ -1,6 +1,6 @@
-/* pepper_mouse_admin: Pepper-mouse spice rodents (v1.0)
+/* pepper_mouse_admin: Pepper-mouse vegetable rodents (v1.0)
  * Pepper-mouse forest, feeding, breeding, health, market
- * Features: body_len_cm, body_wt_g, heat_lv, tail_cm, pepmx_idx, age_year
+ * Features: body_len_cm, body_wt_g, seed_ct, tail_cm, pepmx_idx, age_year
  */
 #include <stddef.h>
 __attribute__((import_module("host"), import_name("alloc"))) extern unsigned int host_alloc(unsigned int, unsigned int);
@@ -8,24 +8,24 @@ __attribute__((import_module("host"), import_name("print"))) extern void host_pr
 __attribute__((import_module("host"), import_name("exit"))) extern void host_exit(int);
 __attribute__((import_module("host"), import_name("get_argv"))) extern int host_get_argv(unsigned int, unsigned int);
 #define N 16
-typedef struct{int id,location,bdy_ln,bdy_wt,heat_lv,tail_cm,pepmx_idx,age_yr,active;} pepmx_t;
-typedef struct{int n_forest,n_feed,n_breed,n_health,n_mkt,t_ln,t_wt,t_hl,t_tail,t_pe;} pepmx_state_t;
+typedef struct{int id,location,bdy_ln,bdy_wt,seed_ct,tail_cm,pepmx_idx,age_yr,active;} pepmx_t;
+typedef struct{int n_forest,n_feed,n_breed,n_health,n_mkt,t_ln,t_wt,t_sc,t_tail,t_pei;} pepmx_state_t;
 static pepmx_t svl[N],svf[N-2],svb[N-4],svh[N-6],svm[N-6]; static pepmx_state_t st; static int init;
 static void ps(const char*s){host_print(s);} static void pi(int v){char b[32];int i=0;if(v<0){b[i++]='-';v=-v;}if(v==0){b[i++]='0';}else{int s=i;while(v>0){b[i++]='0'+(v%10);v/=10;}int e=i-1;while(s<e){char t=b[s];b[s]=b[e];b[e]=t;s++;e--;}}b[i]='\0';host_print(b);}
-static int add(pepmx_t*a,int*cnt,int*sum,int mx,int lc,int bl,int bw,int hl,int tc,int pe,int ay){if(*cnt>=mx)return -1;pepmx_t*x=&a[*cnt];x->id=*cnt;x->location=lc;x->bdy_ln=bl;x->bdy_wt=bw;x->heat_lv=hl;x->tail_cm=tc;x->pepmx_idx=pe;x->age_yr=ay;x->active=1;*sum+=bl;(*cnt)++;ps("[PEPMX] Pepper-mouse ");pi(*cnt-1);ps(" lc=");pi(lc);ps(" bl=");pi(bl);ps(" bw=");pi(bw);ps(" hl=");pi(hl);ps(" tc=");pi(tc);ps(" pe=");pi(pe);ps(" ay=");pi(ay);ps("\n");return *cnt-1;}
-int pepper_mouse_init(void){if(init)return -1;st.n_forest=0;st.n_feed=0;st.n_breed=0;st.n_health=0;st.n_mkt=0;st.t_ln=0;st.t_wt=0;st.t_hl=0;st.t_tail=0;st.t_pe=0;for(int i=0;i<N;i++)svl[i].active=0;for(int i=0;i<N-2;i++)svf[i].active=0;for(int i=0;i<N-4;i++)svb[i].active=0;for(int i=0;i<N-6;i++)svh[i].active=0;for(int i=0;i<N-6;i++)svm[i].active=0;init=1;ps("[PEPMX] Pepper-mouse initialized\n");return 0;}
-int pepper_mouse_forest(int lc,int bl,int bw,int hl,int tc,int pe,int ay){return add(svl,&st.n_forest,&st.t_ln,N,lc,bl,bw,hl,tc,pe,ay);}
-int pepper_mouse_feeding(int lc,int bl,int bw,int hl,int tc,int pe,int ay){return add(svf,&st.n_feed,&st.t_wt,N-2,lc,bl,bw,hl,tc,pe,ay);}
-int pepper_mouse_breeding(int lc,int bl,int bw,int hl,int tc,int pe,int ay){return add(svb,&st.n_breed,&st.t_hl,N-4,lc,bl,bw,hl,tc,pe,ay);}
-int pepper_mouse_health(int lc,int bl,int bw,int hl,int tc,int pe,int ay){return add(svh,&st.n_health,&st.t_tail,N-6,lc,bl,bw,hl,tc,pe,ay);}
-int pepper_mouse_market(int lc,int bl,int bw,int hl,int tc,int pe,int ay){return add(svm,&st.n_mkt,&st.t_pe,N-6,lc,bl,bw,hl,tc,pe,ay);}
-void pepper_mouse_report(void){ps("[PEPMX] Forest: ");pi(st.n_forest);ps(" Ln=");pi(st.t_ln);ps("\nFeed: ");pi(st.n_feed);ps(" Wt=");pi(st.t_wt);ps("\nBreed: ");pi(st.n_breed);ps(" Hl=");pi(st.t_hl);ps("\nHealth: ");pi(st.n_health);ps(" Tail=");pi(st.t_tail);ps("\nMkt: ");pi(st.n_mkt);ps(" Pe=");pi(st.t_pe);ps("\n");}
+static int add(pepmx_t*a,int*cnt,int*sum,int mx,int lc,int bl,int bw,int sc,int tc,int pei,int ay){if(*cnt>=mx)return -1;pepmx_t*x=&a[*cnt];x->id=*cnt;x->location=lc;x->bdy_ln=bl;x->bdy_wt=bw;x->seed_ct=sc;x->tail_cm=tc;x->pepmx_idx=pei;x->age_yr=ay;x->active=1;*sum+=bl;(*cnt)++;ps("[PEPMX] Pepper-mouse ");pi(*cnt-1);ps(" lc=");pi(lc);ps(" bl=");pi(bl);ps(" bw=");pi(bw);ps(" sc=");pi(sc);ps(" tc=");pi(tc);ps(" pei=");pi(pei);ps(" ay=");pi(ay);ps("\n");return *cnt-1;}
+int pepper_mouse_init(void){if(init)return -1;st.n_forest=0;st.n_feed=0;st.n_breed=0;st.n_health=0;st.n_mkt=0;st.t_ln=0;st.t_wt=0;st.t_sc=0;st.t_tail=0;st.t_pei=0;for(int i=0;i<N;i++)svl[i].active=0;for(int i=0;i<N-2;i++)svf[i].active=0;for(int i=0;i<N-4;i++)svb[i].active=0;for(int i=0;i<N-6;i++)svh[i].active=0;for(int i=0;i<N-6;i++)svm[i].active=0;init=1;ps("[PEPMX] Pepper-mouse initialized\n");return 0;}
+int pepper_mouse_forest(int lc,int bl,int bw,int sc,int tc,int pei,int ay){return add(svl,&st.n_forest,&st.t_ln,N,lc,bl,bw,sc,tc,pei,ay);}
+int pepper_mouse_feeding(int lc,int bl,int bw,int sc,int tc,int pei,int ay){return add(svf,&st.n_feed,&st.t_wt,N-2,lc,bl,bw,sc,tc,pei,ay);}
+int pepper_mouse_breeding(int lc,int bl,int bw,int sc,int tc,int pei,int ay){return add(svb,&st.n_breed,&st.t_sc,N-4,lc,bl,bw,sc,tc,pei,ay);}
+int pepper_mouse_health(int lc,int bl,int bw,int sc,int tc,int pei,int ay){return add(svh,&st.n_health,&st.t_tail,N-6,lc,bl,bw,sc,tc,pei,ay);}
+int pepper_mouse_market(int lc,int bl,int bw,int sc,int tc,int pei,int ay){return add(svm,&st.n_mkt,&st.t_pei,N-6,lc,bl,bw,sc,tc,pei,ay);}
+void pepper_mouse_report(void){ps("[PEPMX] Forest: ");pi(st.n_forest);ps(" Ln=");pi(st.t_ln);ps("\nFeed: ");pi(st.n_feed);ps(" Wt=");pi(st.t_wt);ps("\nBreed: ");pi(st.n_breed);ps(" Sc=");pi(st.t_sc);ps("\nHealth: ");pi(st.n_health);ps(" Tail=");pi(st.t_tail);ps("\nMkt: ");pi(st.n_mkt);ps(" Pei=");pi(st.t_pei);ps("\n");}
 void pepper_mouse_state(void){ps("[PEPMX] Forest=");pi(st.n_forest);ps(" Feed=");pi(st.n_feed);ps(" Breed=");pi(st.n_breed);ps(" Health=");pi(st.n_health);ps(" Mkt=");pi(st.n_mkt);ps("\n");}
 int main(void){
 ps("=== Pepper-Mouse Admin Demo ===\n\n");pepper_mouse_init();
-ps("Pepper-mouse forest...\n");for(int i=0;i<N;i++){int lc=(i%5)+1,bl=10+(i*1),bw=18+(i*3),hl=1+(i%10),tc=8+(i*1),pe=(i%7)+1,ay=(i%5)+1;pepper_mouse_forest(lc,bl,bw,hl,tc,pe,ay);}
-ps("\nPepper-mouse feeding...\n");for(int i=0;i<N-2;i++){int lc=(i%4)+2,bl=11+(i*1),bw=20+(i*3),hl=1+(i%10),tc=8+(i*1),pe=(i%6)+1,ay=(i%4)+1;pepper_mouse_feeding(lc,bl,bw,hl,tc,pe,ay);}
-ps("\nPepper-mouse breeding...\n");for(int i=0;i<N-4;i++){int lc=(i%3)+1,bl=12+(i*1),bw=22+(i*3),hl=2+(i%10),tc=9+(i*1),pe=(i%5)+1,ay=(i%3)+1;pepper_mouse_breeding(lc,bl,bw,hl,tc,pe,ay);}
-ps("\nPepper-mouse health...\n");for(int i=0;i<N-6;i++){int lc=(i%5)+1,bl=9+(i*1),bw=16+(i*3),hl=1+(i%10),tc=7+(i*1),pe=(i%9)+1,ay=(i%5)+1;pepper_mouse_health(lc,bl,bw,hl,tc,pe,ay);}
-ps("\nPepper-mouse market...\n");for(int i=0;i<N-6;i++){int lc=(i%4)+1,bl=13+(i*1),bw=24+(i*3),hl=3+(i%10),tc=9+(i*1),pe=(i%4)+1,ay=(i%3)+1;pepper_mouse_market(lc,bl,bw,hl,tc,pe,ay);}
+ps("Pepper-mouse forest...\n");for(int i=0;i<N;i++){int lc=(i%5)+1,bl=10+(i*1),bw=18+(i*3),sc=4+(i%10),tc=8+(i*1),pei=(i%7)+1,ay=(i%5)+1;pepper_mouse_forest(lc,bl,bw,sc,tc,pei,ay);}
+ps("\nPepper-mouse feeding...\n");for(int i=0;i<N-2;i++){int lc=(i%4)+2,bl=11+(i*1),bw=20+(i*3),sc=4+(i%10),tc=8+(i*1),pei=(i%6)+1,ay=(i%4)+1;pepper_mouse_feeding(lc,bl,bw,sc,tc,pei,ay);}
+ps("\nPepper-mouse breeding...\n");for(int i=0;i<N-4;i++){int lc=(i%3)+1,bl=12+(i*1),bw=22+(i*3),sc=5+(i%10),tc=9+(i*1),pei=(i%5)+1,ay=(i%3)+1;pepper_mouse_breeding(lc,bl,bw,sc,tc,pei,ay);}
+ps("\nPepper-mouse health...\n");for(int i=0;i<N-6;i++){int lc=(i%5)+1,bl=9+(i*1),bw=16+(i*3),sc=3+(i%10),tc=7+(i*1),pei=(i%9)+1,ay=(i%5)+1;pepper_mouse_health(lc,bl,bw,sc,tc,pei,ay);}
+ps("\nPepper-mouse market...\n");for(int i=0;i<N-6;i++){int lc=(i%4)+1,bl=13+(i*1),bw=24+(i*3),sc=6+(i%10),tc=9+(i*1),pei=(i%4)+1,ay=(i%3)+1;pepper_mouse_market(lc,bl,bw,sc,tc,pei,ay);}
 ps("\n");pepper_mouse_report();pepper_mouse_state();ps("\n=== Demo Complete ===\n");return 0;}